requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.0.7
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.7
incremental==24.7.1
msgpack==1.0.8
pillow==10.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
redis==5.0.7
service-identity==24.1.0
setuptools==71.1.0
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
zope.interface==6.4.post2
